[optimization] Use RWLock Hash with Mutex Queue in connection pool
The original data structure is a hash using MUTEX to guarantee thread
safety. It uses `server_name#server_version` as key, connection instances
as value. Usually, a key corresponds to multiple values, so it's a
multi-map. Because the connection pool is similar to the
consumer-producer model, we redesigned the data structure, using a
unique hash map, where its key is `server_name#server_version`, value is
a queue of connection instances.
We use RWLock to secure the hash map since we don't usually change the
server version, the vast majority of operations are search, rather than
insertion or deletion. Therefore, an RWLock will boost performance
significantly. We use a mutex to secure the queue. So we convert a
single mutex to multiple mutexes, which can also increase concurrency.