[optimization] Disable bgs_mode when a SELECT fetches the whole table
When a SELECT can get a huge load of results (like when fetching the whole table), propagating queries to remotes *in parallel* may lead to feeding Spider too much for it to process at once, which can directly cause network timeout. Parallel propagation of queries should be disabled under this circumstance.

For this commit, we allow a SELECT query to be propagated in parallel only when:
1. It contains aggregate functions.
2. It contains GROUP BY, or LIMIT clauses.
3. It contains a WHERE cond.

P.S. This commit is more of a temporary solution, the decision should be made with more factors being considered (maybe by utilizing mysql.spider_table_status).