[bugfix] Fix connection leaks in spd_connect_pools
Before this commit, the hash_key of conn queues is a pointer ref to the conn_key of the first conn pushed to the queue. This could cause conn object leaks if the first conn gets freed (due to either the recycle process or a query execution error), while some other conns with the same conn_key just get pushed into the queue. Then the queue's hash_key is messed up and the conns in the queue will be lost forever, causing the corresponding ip_port_conn->ip_port_count to stack up and reach SPIDER_MAX_CONNECTIONS, then the data node is no longer accessible.

Solution: alloc memory for cq->hash_key and copy the conn_key into it.